requirements.txt
# FastMCP framework
fastmcp==2.13.3

# FastAPI & ASGI server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# HTTP client
httpx>=0.27.0

# JWT token handling
PyJWT>=2.8.0

# Data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# SSE support
sse-starlette>=2.0.0
